+++ src/cli.ts
@@ -1,4 +1,3 @@
-#!/usr/bin/env node
 import { runCli } from './run.js'
 
 runCli(process.argv.slice(2), {